agendamento_service/requirements.txt
psycopg2-binary
grpcio
grpcio-tools